simplify the use of send and quote
Methods quote and send was a bit confusing, even after reading the
readme. Refactor so that send is the raw sender, and add a command
utility to send commands with auto colon added to last arg.
Additionally made the msg other friendly functions accept a simple
msg arg, as it is very unlikely to need to pass multiple params to
those methods.